+++ src/styles/cards.css
@@ -1,0 +1,125 @@
+
+.cards-text {
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 800;
+  font-size: 26px;
+  line-height: 40px;
+  text-align: center;
+  color: var(--primary);
+  margin: auto;
+}
+
+.cards {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--gap-cards);
+  width: 100%;
+}
+
+.card {
+  display: flex;
+  flex-direction: column;
+  width: calc((100% - var(--gap-cards) * (var(--cards-in-row) * var(--cards-multiplier) - 1)) / (var(--cards-in-row) * var(--cards-multiplier)));
+  justify-items: center;
+  align-items: center;
+  background-color: #F5F5F5;
+  padding: 10px;
+  border-radius: 5px;
+}
+
+.card__discount {
+  align-self: flex-end;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 64px;
+  height: 29px;
+  background: #E56B50;
+  border-radius: 10px;
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 14px;
+  line-height: 18px;  
+  color: #FFFFFF;
+}
+
+.card__title {  
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 800;
+  font-size: 16px;
+  line-height: 21px;
+  color: #009FDF;
+}
+
+.card__image {
+  width: 100%;
+  aspect-ratio: 1/1;
+  background-repeat: no-repeat;  
+  background-size: cover;
+}
+
+.card__price {
+  align-self: flex-start;
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 22px;
+  line-height: 28px;
+}
+
+.card__btns {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+}
+
+.card__add-btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10%;
+  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
+  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
+  border-radius: 5px;
+  border: none;
+  width: 48%;
+  height: 40px;
+}
+
+.card__add-btn:hover {
+  background: linear-gradient(0deg, rgba(0, 0, 0, 0.15), rgba(0, 0, 0, 0.15)), linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
+}
+
+.add-to-cart {
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 13px;
+  line-height: 16px;
+  text-transform: uppercase;
+  color: white;
+}
+
+.card__detail-btn {
+  background: #F5F5F5;
+  border-radius: 5px;
+  border: 2px solid #009FDF;
+  width: 48%;
+  height: 40px;
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 13px;
+  line-height: 16px;
+  text-transform: uppercase;
+  color: #009FDF;
+  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
+}
+
+.card__detail-btn:hover {
+  background: #009FDF;
+  color: #F5F5F5;
+}
