+++ src/styles/checkboxes.css
@@ -1,0 +1,71 @@
+.checkbox-line {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.checkbox-line .label {
+  margin-right: auto;
+}
+
+.checkbox {
+  margin: 0 5px;
+  opacity: 0;
+  cursor: pointer;
+}
+
+.label {
+  position: relative;
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 14px;
+  line-height: 21px;
+  text-transform: capitalize;
+}
+
+.checkmark {
+  position: absolute;
+  top: 0;
+  left: -22px;
+  height: 20px;
+  width: 20px;
+  background: #FFFFFF;
+  border-radius: 50%;
+  border: 1px solid #6DA5C6;
+  cursor: pointer;
+}
+
+input:checked ~ .label {
+  color: var(--primary);
+}
+
+input:hover ~ .label .checkmark {
+  background-color: #ededed;
+}
+
+.checkmark:after {
+  content: "";
+  position: absolute;
+  display: none;
+}
+
+input:checked ~ .label .checkmark:after {
+  display: block;
+}
+
+.checkmark:after {
+  left: 6px;
+  top: -5px;
+  width: 7px;
+  height: 17px;
+  border: solid var(--primary);
+  border-width: 0 3px 3px 0;
+  -webkit-transform: rotate(45deg);
+  -ms-transform: rotate(45deg);
+  transform: rotate(45deg);
+}
+
+.brand-count {
+  margin-right: 5px;
+}
