+++ src/styles/products.css
@@ -3,16 +3,7 @@
   --cards-in-row: 4;
   --cards-multiplier: 1;
   --btn-width: 48%;
-  --primary-light: #8abdff;
-  /* --primary: #6d5dfc; */
   --primary: #2196F3;
-  --primary-dark: #5b0eeb;
-  --white: #ffffff;
-  --greyLight-1: #e4ebf5;
-  --greyLight-2: #c8d0e7;
-  --greyLight-3: #bec8e4;
-  --greyDark: #9baacf;
-
 }
 
 * {
@@ -27,398 +18,6 @@ button {
   max-width: 1200px;
   margin: 0 auto;
   padding: 0 10px;
-}
-
-.filter {
-  display: flex;
-  gap: var(--gap-cards);
-  background: #FFFFFF;
-  box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.25);
-  border-radius: 10px;
-  margin: 20px 0;
-  width: 100%;
-}
-
-/* .filter__brand,
-.filter__category, */
-.filter__block {
-  display: flex;
-  flex-direction: column;
-  width: calc((100% - var(--gap-cards) * 2) / 3);
-  padding: 10px;
-}
-
-/* .filter__brand-header,
-.filter__category-header, */
-.filter__header {
-  text-align: center;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 18px;
-  line-height: 23px;
-  text-transform: uppercase;
-  color: #000000;
-}
-
-.filter__list {
-  height: 300px;
-  overflow-y: scroll;
-}
-
-.filter__list::-webkit-scrollbar-thumb {
-  background: #D9D9D9;
-  border-radius: 3px;
-}
-
-.filter__list::-webkit-scrollbar-thumb:hover {
-  background: linear-gradient(0deg, rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.2)), #D9D9D9; 
-}
-
-.filter__list::-webkit-scrollbar {
-  width: 10px;
-  background: #ededed;
-  border-radius: 3px;
-}
-
-
-
-
-.range__container {
-  display: flex;
-  flex-direction: column;
-  width: 80%;
-  margin: 5% auto;
-}
-
-.range__title {
-  margin-bottom: 1em;
-}
-
-.range__sliders {
-  position: relative;
-  min-height: 15px;
-}
-
-input[type=range]::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  pointer-events: all;
-  width: 14px;
-  height: 24px;
-  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);;
-  border-radius: 5px;
-  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
-  cursor: pointer;
-}
-
-input[type=range]::-moz-range-thumb {
-  --webkit-appearance: none;
-  pointer-events: all;
-  width: 14px;
-  height: 24px;
-  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);;
-  border-radius: 5px;
-  box-shadow: 0 0 0 1px #C6C6C6;
-  cursor: pointer;  
-}
-
-input[type=range]::-webkit-slider-thumb:hover {
-  background: linear-gradient(0deg, rgba(0, 0, 0, 0.15), rgba(0, 0, 0, 0.15)), linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
-}
-
-input[type=range]::-webkit-slider-thumb:active {
-  box-shadow: inset 0 0 3px #387bbe, 0 0 9px #387bbe;
-  -webkit-box-shadow: inset 0 0 3px #387bbe, 0 0 9px #387bbe;
-}
-
-input[type="range"] {
-  -webkit-appearance: none; 
-  appearance: none;
-  height: 2px;
-  width: 100%;
-  position: absolute;
-  background-color: #C6C6C6;
-  pointer-events: none;
-}
-
-#fromPrice,
-#fromStock {
-  height: 0;
-  z-index: 1;
-}
-
-.filter__values {
-  display: flex;
-  justify-content: space-between;
-}
-
-.filter__value {
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 18px;
-}
-
-
-
-
-.checkbox-line {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.checkbox-line .label {
-  margin-right: auto;
-}
-
-.checkbox {
-  margin: 0 5px;
-  opacity: 0;
-  cursor: pointer;
-}
-
-.brand-count {
-  margin-right: 5px;
-}
-
-.label {
-  position: relative;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 21px;
-  text-transform: capitalize;
-}
-
-.checkmark {
-  position: absolute;
-  top: 0;
-  left: -22px;
-  height: 20px;
-  width: 20px;
-  background: #FFFFFF;
-  border-radius: 50%;
-  border: 1px solid #6DA5C6;
-  cursor: pointer;
-}
-
-input:checked ~ .label {
-  color: var(--primary);
-}
-
-input:hover ~ .label .checkmark {
-  background-color: #ededed;
-}
-
-.checkmark:after {
-  content: "";
-  position: absolute;
-  display: none;
-}
-
-input:checked ~ .label .checkmark:after {
-  display: block;
-}
-
-.checkmark:after {
-  left: 6px;
-  top: -5px;
-  width: 7px;
-  height: 17px;
-  border: solid var(--primary);
-  border-width: 0 3px 3px 0;
-  -webkit-transform: rotate(45deg);
-  -ms-transform: rotate(45deg);
-  transform: rotate(45deg);
-}
-
-
-.settings {
-  width: 100%;
-  margin: 1rem 0;
-  height: 4rem;
-  box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.25);
-  border-radius: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  /* position: relative; */
-}
-
-.settings__sort,
-.settings__size {
-  display: flex;
-  align-items: center;
-  position: relative;
-  margin: 1em;
-}
-
-.settings__size input,
-.settings__sort input {
-  display: none;
-}
-
-.settings__size > input:checked + label,
-.settings__sort > input:checked + label {
-  color: var(--primary);
-  border:  1px solid #6DA5C6;
-  border-radius: 5px;
-  /* transition: all 0.5s ease; */
-}
-
-.size-control,
-.sort-control {
-  width: 6.0rem;
-  height: 2.6rem;
-  font-size: 1em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  transition: all 0.5s ease;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 500;
-  font-size: 13px;
-  line-height: 16px;
-  color: #495955;
-}
-
-.size-control {
-  width: 3.0rem;
-}
-
-.size-control:hover,
-.sort-control:hover {
-  color: var(--primary);
-}
-
-.cards-text {
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 800;
-  font-size: 26px;
-  line-height: 40px;
-  text-align: center;
-  color: var(--primary);
-  margin: auto;
-}
-
-.cards {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--gap-cards);
-  width: 100%;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  width: calc((100% - var(--gap-cards) * (var(--cards-in-row) * var(--cards-multiplier) - 1)) / (var(--cards-in-row) * var(--cards-multiplier)));
-  justify-items: center;
-  align-items: center;
-  background-color: #F5F5F5;
-  padding: 10px;
-  border-radius: 5px;
-}
-
-.card__discount {
-  align-self: flex-end;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 64px;
-  height: 29px;
-  background: #E56B50;
-  border-radius: 10px;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 18px;  
-  color: #FFFFFF;
-}
-
-.card__title {  
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 800;
-  font-size: 16px;
-  line-height: 21px;
-  color: #009FDF;
-}
-
-.card__image {
-  width: 100%;
-  aspect-ratio: 1/1;
-  background-repeat: no-repeat;  
-  background-size: cover;
-}
-
-.card__price {
-  align-self: flex-start;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 22px;
-  line-height: 28px;
-}
-
-.card__btns {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-}
-
-.card__add-btn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10%;
-  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
-  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
-  border-radius: 5px;
-  border: none;
-  width: 48%;
-  height: 40px;
-}
-
-.card__add-btn:hover {
-  background: linear-gradient(0deg, rgba(0, 0, 0, 0.15), rgba(0, 0, 0, 0.15)), linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
-}
-
-.add-to-cart {
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 16px;
-  text-transform: uppercase;
-  color: white;
-}
-
-.card__detail-btn {
-  background: #F5F5F5;
-  border-radius: 5px;
-  border: 2px solid #009FDF;
-  width: 48%;
-  height: 40px;
-  font-family: 'Geometria';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 13px;
-  line-height: 16px;
-  text-transform: uppercase;
-  color: #009FDF;
-  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
-}
-
-.card__detail-btn:hover {
-  background: #009FDF;
-  color: #F5F5F5;
 }
 
 @media (max-width: 1079px) {
