+++ src/styles/ranges.css
@@ -1,0 +1,62 @@
+.range__container {
+  display: flex;
+  flex-direction: column;
+  width: 90%;
+  margin: 0 auto 5%;
+}
+
+.range__title {
+  margin-bottom: 1em;
+}
+
+.range__sliders {
+  position: relative;
+  min-height: 15px;
+} 
+
+input[type=range]::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  pointer-events: all;
+  width: 14px;
+  height: 24px;
+  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);;
+  border-radius: 5px;
+  box-shadow: 0px 2px 10px 1px rgba(0, 0, 0, 0.15);
+  cursor: pointer;
+}
+
+input[type=range]::-moz-range-thumb {
+  --webkit-appearance: none;
+  pointer-events: all;
+  width: 14px;
+  height: 24px;
+  background: linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);;
+  border-radius: 5px;
+  box-shadow: 0 0 0 1px #C6C6C6;
+  cursor: pointer;  
+}
+
+input[type=range]::-webkit-slider-thumb:hover {
+  background: linear-gradient(0deg, rgba(0, 0, 0, 0.15), rgba(0, 0, 0, 0.15)), linear-gradient(90deg, #0099DC 7.09%, #3CC39D 93.66%);
+}
+
+input[type=range]::-webkit-slider-thumb:active {
+  box-shadow: inset 0 0 3px #387bbe, 0 0 9px #387bbe;
+  -webkit-box-shadow: inset 0 0 3px #387bbe, 0 0 9px #387bbe;
+}
+
+input[type="range"] {
+  -webkit-appearance: none; 
+  appearance: none;
+  height: 2px;
+  width: 100%;
+  position: absolute;
+  background-color: #C6C6C6;
+  pointer-events: none;
+}
+
+#fromPrice,
+#fromStock {
+  height: 0;
+  z-index: 1;
+}
