+++ src/styles/settings.css
@@ -1,0 +1,66 @@
+.settings__found {
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 22px;
+  color: #495955;
+}
+
+.settings {
+  width: 100%;
+  margin: 1rem 0;
+  height: 4rem;
+  box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.25);
+  border-radius: 10px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.settings__sort,
+.settings__size {
+  display: flex;
+  align-items: center;
+  position: relative;
+  margin: 1em;
+}
+
+.settings__size input,
+.settings__sort input {
+  display: none;
+}
+
+.settings__size > input:checked + label,
+.settings__sort > input:checked + label {
+  color: var(--primary);
+  border:  1px solid #6DA5C6;
+  border-radius: 5px;
+}
+
+.size-control,
+.sort-control {
+  width: 6.0rem;
+  height: 2.6rem;
+  font-size: 1em;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  transition: all 0.5s ease;
+  font-family: 'Geometria';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 13px;
+  line-height: 16px;
+  color: #495955;
+}
+
+.size-control {
+  width: 3.0rem;
+}
+
+.size-control:hover,
+.sort-control:hover {
+  color: var(--primary);
+}
